+++ ndonnx/_array.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo 2023-2025
+# Copyright (c) QuantCo 2023-2026
 # SPDX-License-Identifier: BSD-3-Clause
 
 from __future__ import annotations
@@ -89,7 +89,6 @@ def _make_binary_dunder(
     backward_name: str,
 ) -> tuple[_BinaryOp, _BinaryOp]:
     """Create a forward and reflected version for a binary dunder method."""
-
     # If we return 'NotImplemented' from methods such as __add__ the
     # interpreter will create an error message that does not display
     # the arrays dtype. E.g. `"TypeError: ... +: Not Implemented for
@@ -212,11 +211,9 @@ class Array:
         """
         # Special cases that allow for shortcuts
         if self.ndim == 0:
-            Array._from_tyarray(onnx.const(1, dtype=onnx.int64))
-        if self.ndim == 1:
-            return self.dynamic_shape[0]
-
-        return Array._from_tyarray(self.dynamic_shape._tyarray.prod())
+            return Array._from_tyarray(onnx.const(1, dtype=onnx.int64))
+        size = self._tyarray.dynamic_size
+        return Array._from_tyarray(size)
 
     @property
     def T(self) -> Array:  # noqa: N802
@@ -516,11 +513,13 @@ def _normalize_setitem_key_item(
         if isinstance(el, int | None):
             return el
         if not isinstance(el.dtype, onnx.Integer):
-            IndexError(
+            raise IndexError(
                 f"arrays in 'slice' objects must be of integer data types; found `{el.dtype}"
             )
         if el.ndim != 0:
-            IndexError(f"arrays in 'slice' objects must be rank-0; found `{el.ndim}")
+            raise IndexError(
+                f"arrays in 'slice' objects must be rank-0; found `{el.ndim}"
+            )
         return el._tyarray.astype(onnx.int64)
 
     if isinstance(item, slice):