+++ ndonnx/_dtypes.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo 2023-2025
+# Copyright (c) QuantCo 2023-2026
 # SPDX-License-Identifier: BSD-3-Clause
 
 from __future__ import annotations
@@ -37,8 +37,8 @@ class DType(ABC, Generic[TY_ARRAY_BASE]):
     def __ndx_cast_from__(self, arr: TyArrayBase) -> TY_ARRAY_BASE:
         """Convert the given array to this data type.
 
-        This function is used to implement ``TyArrayBase.astype`` and
-        should not be called directly.
+        This function is used to implement ``TyArrayBase.astype`` and should not be
+        called directly.
         """
         ...