+++ ndonnx/_typed_array/onnx.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo 2023-2025
+# Copyright (c) QuantCo 2023-2026
 # SPDX-License-Identifier: BSD-3-Clause
 
 from __future__ import annotations
@@ -35,18 +35,22 @@ from .indexing import FancySlice
 
 _ScalarInt: TypeAlias = "TyArrayInteger"
 """Alias signaling that this must be a rank-0 integer tensor."""
+
 _BoolMask: TypeAlias = "TyArrayBool"
 """Alias signaling that this must be a rank-1 boolean tensor."""
+
 SetitemItem: TypeAlias = "int | slice | EllipsisType | _ScalarInt"
 """A single item; i.e. not a tuple nor a boolean mask.
 
 This does not include `None`.
 """
+
 GetitemItem: TypeAlias = "int | slice | EllipsisType | _ScalarInt | None"
 """A single item (; i.e. not a tuple nor a boolean mask) for __getitem__.
 
 This includes `None`.
 """
+
 SetitemIndex: TypeAlias = "SetitemItem | tuple[SetitemItem, ...] | _BoolMask"
 GetitemIndex: TypeAlias = "GetitemItem | tuple[GetitemItem, ...] | _BoolMask"
 
@@ -71,12 +75,11 @@ def _inline(
 ) -> Callable[Concatenate[TY_ARRAY, P], TY_ARRAY_OUT]:
     """Build the wrapped function as a self-contained ONNX graph and inline it.
 
-    This is useful for functions which have to use the `If` operator
-    in order to work around bugs in the onnxruntime. Without this
-    wrapper, value propagation will be executed in either arm of the
-    `If` node and subsequently fail in one of them (it is the point of
-    the `If` node to avoid the computation of the problematic branch
-    at inference time).
+    This is useful for functions which have to use the `If` operator in order to work
+    around bugs in the onnxruntime. Without this wrapper, value propagation will be
+    executed in either arm of the `If` node and subsequently fail in one of them (it is
+    the point of the `If` node to avoid the computation of the problematic branch at
+    inference time).
     """
 
     @wraps(fun)
@@ -213,7 +216,7 @@ class _Number(_OnnxDType[TY_ARRAY_co]):
         # large numbers. See hypothesis test examples.
 
         if all(isinstance(el, int | float) for el in [start, stop, step]):
-            return const(np.arange(start, stop, step)).astype(self)
+            return const(np.arange(start, stop, step)).astype(self)  # type: ignore
         sss = [
             const(el, self) if isinstance(el, int | float) else el.astype(self)
             for el in [start, stop, step]
@@ -570,10 +573,10 @@ class TyArray(TyArrayBase):
                 key = key[tuple(idx)]
             self._var = op.where(key._var, value._var, self._var)
             return
-        elif value.ndim != 1:
-            # NumPy semantics
-            TypeError(
-                f"assignment value must be 0 or 1-dimensional for boolean indexing, got `{value.ndim}`"
+        if key.ndim == self.ndim and value.ndim > 1:
+            raise TypeError(
+                "ndonnx boolean array indexing assignment requires a 0 or "
+                f"1-dimensional input, input has {value.ndim} dimensions"
             )
 
         # The following is essentially doing `self[ndx.nonzero(key)] = value`
@@ -1463,6 +1466,15 @@ class TyArrayNumber(TyArray):
     def __radd__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
         return self._apply(other, op.add, forward=False, result_type=TyArrayNumber)
 
+    def __rfloordiv__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
+        # See comment in TyArraySignedInteger.__floordiv__ for further
+        # context and why this may be a good example for the `__r*__`
+        # methods.
+        if isinstance(other, int | float):
+            rhs, lhs = promote(self, other)
+            return lhs.__floordiv__(rhs)
+        return NotImplemented
+
     @overload
     def __ge__(self, other: TyArrayNumber | int | float) -> TyArrayBool: ...
 
@@ -1525,6 +1537,12 @@ class TyArrayNumber(TyArray):
     def __rmul__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
         return self._apply(other, op.mul, forward=False, result_type=TyArrayNumber)
 
+    def __rmod__(self, other) -> TyArrayBase:
+        if isinstance(other, int | float):
+            b, a = promote(self, other)
+            return a.__mod__(b)
+        return NotImplemented
+
     def __pow__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
         return self._apply(other, op.pow, forward=True, result_type=TyArrayNumber)
 
@@ -1551,18 +1569,6 @@ class TyArrayNumber(TyArray):
 
     def __rtruediv__(self, other) -> TyArrayBase:
         return self._apply(other, op.div, forward=False, result_type=TyArrayNumber)
-
-    def __floordiv__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
-        if isinstance(other, TyArrayNumber | int | float):
-            promo_result, _ = promote(self, other)
-            return (self / other).floor().astype(promo_result.dtype)
-        return NotImplemented
-
-    def __rfloordiv__(self, other) -> TyArrayBase:
-        if isinstance(other, int | float):
-            promo_result, _ = promote(self, other)
-            return (other / self).floor().astype(promo_result.dtype)
-        return NotImplemented
 
     def sign(self) -> Self:
         return type(self)(op.sign(self._var))
@@ -1709,15 +1715,16 @@ class TyArrayInteger(TyArrayNumber):
     def __ror__(self, other) -> TyArrayBase:
         return self._apply_int_only(other, op.bitwise_or, forward=False)
 
-    def __mod__(self, other) -> TyArrayInteger:
-        return self._apply_int_only(
-            other, lambda a, b: op.mod(a, b, fmod=0), forward=True
-        )
+    def __mod__(self, other) -> TyArrayBase:
+        if isinstance(other, type(self)):
+            var = op.mod(self._var, other._var, fmod=0)
+            return safe_cast(type(self), _var_to_tyarray(var))
 
-    def __rmod__(self, other) -> TyArrayInteger:
-        return self._apply_int_only(
-            other, lambda a, b: op.mod(a, b, fmod=0), forward=False
-        )
+        if isinstance(other, TyArrayNumber | int | float):
+            a, b = promote(self, other)
+            return a.__mod__(b)
+
+        return NotImplemented
 
     def __xor__(self, other) -> TyArrayBase:
         return self._apply_int_only(other, op.bitwise_xor, forward=True)
@@ -1789,6 +1796,36 @@ class TyArrayInteger(TyArrayNumber):
 
 
 class TyArraySignedInteger(TyArrayInteger):
+    def __floordiv__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
+        # TODO: At the time of writing I believe this to be a good
+        # general approach to implementing dunder methods. We should
+        # consider combing through the existing one to have them all
+        # on the same footing.
+        #
+        # 1. Select the cases where we know exactly what to do without
+        # further promotion.
+        # 2. Select the cases where we know that a type promotion
+        # should happen (i.e `isinstance(other, int | float | TyArrayNumber)`,
+        # but the ultimate implementation depends on the result of the
+        # promotion.
+        # 3. `return NotImplemented`
+        # 4. `__r*__` only ever deals with python scalars and there it
+        # always must use `promote`. These should actually live in `TyArray`.
+        if isinstance(other, type(self)):
+            # div operator truncates towards zero, but we need to floor
+            trunc = _binary(op.div, type(self))(self, other)
+            needs_fix = safe_cast(
+                TyArrayBool, ((self ^ other) < 0) & ((self % other) != 0)
+            )
+            return trunc - needs_fix.astype(self.dtype)
+
+        if isinstance(other, TyArrayNumber | int | float):
+            # Deferred to the promoted type.
+            a, b = promote(self, other)
+            return a.__floordiv__(b)
+
+        return NotImplemented
+
     # The array-api standard defines the right shift as arithmetic
     # (i.e. sign-propagating). The ONNX standard is logical.
 
@@ -1804,6 +1841,18 @@ class TyArraySignedInteger(TyArrayInteger):
 
 
 class TyArrayUnsignedInteger(TyArrayInteger):
+    def __floordiv__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
+        if isinstance(other, type(self)):
+            # div does truncation, but that is fine if we only deal with unsigned integers
+            return _binary(op.div, type(self))(self, other)
+
+        if isinstance(other, TyArrayNumber | int | float):
+            # Deferred to the promoted type.
+            a, b = promote(self, other)
+            return a.__floordiv__(b)
+
+        return NotImplemented
+
     def __abs__(self) -> Self:
         return self.copy()
 
@@ -1843,34 +1892,41 @@ class TyArrayUnsignedInteger(TyArrayInteger):
 
 
 class TyArrayFloating(TyArrayNumber):
+    def __floordiv__(self, other: TyArrayBase | PyScalar) -> TyArrayBase:
+        if isinstance(other, type(self)):
+            return _binary(op.div, type(self))(self, other).floor()
+
+        if isinstance(other, TyArrayNumber | int | float):
+            a, b = promote(self, other)
+            return a.__floordiv__(b)
+
+        return NotImplemented
+
     def __mod__(self, other) -> TyArrayBase:
-        if isinstance(other, TyArrayFloating | float):
+        if isinstance(other, type(self)):
             # This function is complicated for two reasons:
             # 1. The ONNX standard is undefined if dividend is 0, but the array-api is not.
             # 2. The array-api follows the Python semantics, which are rather odd.
-            a, b = promote(self, other)
-            var = op.mod(a._var, b._var, fmod=1)
+            var = op.mod(self._var, other._var, fmod=1)
             mod = safe_cast(TyArrayFloating, _var_to_tyarray(var))
             # NOTE: onnxruntime appears to have a bug where the sign
             # of zeros is only preserved if they are on the
             # false-branch!
             # TODO: File a bug!
-            fixed_mod = where((b < 0) == (mod < 0), mod, mod + b)
-            fixed_zeros = where(b > 0, const(0.0, mod.dtype), const(-0.0, mod.dtype))
+            fixed_mod = where((other < 0) == (mod < 0), mod, mod + other)
+            fixed_zeros = where(
+                other > 0, const(0.0, mod.dtype), const(-0.0, mod.dtype)
+            )
             return where(
-                safe_cast(TyArrayBool, ~((mod == 0.0) & (b != 0.0))),
+                safe_cast(TyArrayBool, ~((mod == 0.0) & (other != 0.0))),
                 fixed_mod,
                 fixed_zeros,
             )
+        if isinstance(other, TyArrayNumber | int | float):
+            a, b = promote(self, other)
+            return a.__mod__(b)
 
-        return super().__mod__(other)
-
-    def __rmod__(self, other) -> TyArrayBase:
-        if isinstance(other, TyArrayFloating | float):
-            b, a = promote(self, other)
-            var = op.mod(a._var, b._var, fmod=1)
-            return safe_cast(TyArrayFloating, _var_to_tyarray(var))
-        return super().__mod__(other)
+        return NotImplemented
 
     def __ndx_logaddexp__(self, x2: TyArrayBase | int | float, /) -> TyArrayFloating:
         if isinstance(x2, TyArrayNumber | int | float):
@@ -2284,7 +2340,8 @@ def const(
 ) -> TyArray:
     """Create a constant from the given value.
 
-    The `dtype` argument may be used in favor of a subsequent `astype` call to create a cleaner ONNX graph.
+    The `dtype` argument may be used in favor of a subsequent `astype` call to create a
+    cleaner ONNX graph.
     """
     # don't blindly fall back to NumPy to maintain better np1x
     # compatibility on Windows which defaults to int32
@@ -2607,7 +2664,6 @@ def _move_ellipsis_back(
     key: tuple[SetitemItem, ...],
 ) -> tuple[tuple[int, ...], tuple[int, ...], tuple[SetitemItem, ...]]:
     """Permute axes such that the ellipsis-axes are at the end."""
-
     if ... not in key:
         raise ValueError("No ellipsis found in 'key'")
     ellipsis_pos = key.index(...)
@@ -2905,3 +2961,13 @@ def _output_reduce_zero_size(
         )
         out_shape = out_shape.take(const(axes_to_keep, int64))
     return value.broadcast_to(out_shape)
+
+
+def _binary(
+    op: Callable[[Var, Var], Var], otype: type[TY_ARRAY]
+) -> Callable[[TY_ARRAY, TY_ARRAY], TY_ARRAY]:
+    def do(a: TY_ARRAY, b: TY_ARRAY) -> TY_ARRAY:
+        res = op(a._var, b._var)
+        return safe_cast(otype, _var_to_tyarray(res))
+
+    return do