+++ ndonnx/extensions.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo 2023-2025
+# Copyright (c) QuantCo 2023-2026
 # SPDX-License-Identifier: BSD-3-Clause
 
 from __future__ import annotations
@@ -10,9 +10,11 @@ import numpy as np
 from typing_extensions import TypeIs, deprecated
 
 import ndonnx as ndx
-
-from . import _typed_array as tydx
-from ._typed_array.masked_onnx import TyMaArray
+import ndonnx._typed_array as tydx
+import ndonnx._typed_array.datetime
+import ndonnx._typed_array.funcs
+import ndonnx._typed_array.masked_onnx
+import ndonnx._typed_array.onnx
 
 SCALAR = TypeVar("SCALAR", int, float, str)
 
@@ -198,7 +200,7 @@ def make_nullable(
     null = None if null is None else null.copy()
 
     if null is None:
-        if isinstance(x._tyarray, TyMaArray):
+        if isinstance(x._tyarray, tydx.masked_onnx.TyMaArray):
             return x
         if isinstance(x._tyarray, tydx.onnx.TyArray):
             return ndx.Array._from_tyarray(
@@ -248,9 +250,9 @@ def get_data(x: ndx.Array, /) -> ndx.Array:
 def put(a: ndx.Array, indices: ndx.Array, updates: ndx.Array, /) -> None:
     """Replaces specified elements of an array with given values.
 
-    This function follows the semantics of `numpy.put` with
-    `mode="raises". The data types of the update array and the updates
-    must match. The indices must be provided as a 1D int64 array.
+    This function follows the semantics of `numpy.put` with `mode="raises". The data
+    types of the update array and the updates must match. The indices must be provided
+    as a 1D int64 array.
     """
     if not isinstance(indices._tyarray, tydx.onnx.TyArrayInt64):
         # using isinstance here to get the type narrowing below
@@ -399,6 +401,12 @@ def is_nullable_dtype(dtype: ndx.DType, /) -> TypeIs[tydx.masked_onnx.DTypes]:
     return isinstance(dtype, tydx.masked_onnx.DTypes)
 
 
+def is_nullable_numeric_dtype(
+    dtype: ndx.DType, /
+) -> TypeIs[tydx.masked_onnx.NumericDTypes]:
+    return is_nullable_integer_dtype(dtype) or is_nullable_float_dtype(dtype)
+
+
 def is_nullable_integer_dtype(
     dtype: ndx.DType, /
 ) -> TypeIs[tydx.masked_onnx.IntegerDTypes]:
@@ -423,11 +431,13 @@ __all__ = [
     "datetime_to_year_month_day",
     "fill_null",
     "get_mask",
+    "get_data",
     "is_float_dtype",
     "is_integer_dtype",
     "is_nullable_dtype",
     "is_nullable_float_dtype",
     "is_nullable_integer_dtype",
+    "is_nullable_numeric_dtype",
     "is_numeric_dtype",
     "is_onnx_dtype",
     "is_signed_integer_dtype",